code/requirements.txt
python-dotenv
google-genai
requests
beautifulsoup4
textual